# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Store API Documentation",
        "description": "Store API Postman and Code definitions documentation.",
        "contact": {
            "name": "Backend Engineer"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/admin/orders": {
            "get": {
                "tags": [
                    "Admin Orders"
                ],
                "summary": "Admin List Orders",
                "description": "List all orders with filtering",
                "operationId": "5d3d4445b82b0c1e78a9494d4146a376",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "payment_status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orders fetched",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Order"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/orders/{id}": {
            "get": {
                "tags": [
                    "Admin Orders"
                ],
                "summary": "Admin Show Order",
                "description": "Show details of a specific order",
                "operationId": "4fb50d5065b6b7251972d6adbf396e1b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order fetched",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/orders/{id}/status": {
            "post": {
                "tags": [
                    "Admin Orders"
                ],
                "summary": "Admin Update Order Status",
                "description": "Transition order status or payment status",
                "operationId": "90f2c8b7e361a200f6315571200ddbc3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "description": "pending, processing, shipped, delivered, cancelled",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "payment_status": {
                                        "description": "unpaid, paid, failed, refunded",
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order status updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Status transition not allowed"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/orders/{order}/refund": {
            "post": {
                "tags": [
                    "Admin Orders"
                ],
                "summary": "Admin Refund Order",
                "description": "Issue a full Stripe refund for a paid order and mark it as refunded.",
                "operationId": "c0abfd1268155e28733e289facc942bb",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order refunded"
                    },
                    "409": {
                        "description": "Order not eligible for refund"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "NotFoundResponse": {
                "title": "NotFound Response",
                "description": "Response for 404 not found error",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Resource not found."
                    },
                    "data": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "errors": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UnauthorizedResponse": {
                "title": "Unauthorized Response",
                "description": "Response for 401 unauthorized error",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    },
                    "data": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "errors": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Address": {
                "title": "Address",
                "description": "User Address schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Home"
                    },
                    "address_line_1": {
                        "type": "string",
                        "example": "123 Main St"
                    },
                    "address_line_2": {
                        "type": "string",
                        "example": "Apt 4B",
                        "nullable": true
                    },
                    "city": {
                        "type": "string",
                        "example": "New York"
                    },
                    "state": {
                        "type": "string",
                        "example": "NY"
                    },
                    "postal_code": {
                        "type": "string",
                        "example": "10001"
                    },
                    "country": {
                        "type": "string",
                        "example": "USA"
                    },
                    "phone": {
                        "type": "string",
                        "example": "123-456-7890"
                    },
                    "is_default": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Order": {
                "title": "Order",
                "description": "Order schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "order_number": {
                        "type": "string",
                        "example": "ORD-2023-1001"
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "payment_status": {
                        "type": "string",
                        "example": "paid"
                    },
                    "subtotal": {
                        "type": "number",
                        "format": "float",
                        "example": 299.99
                    },
                    "tax": {
                        "type": "number",
                        "format": "float",
                        "example": 0
                    },
                    "shipping_cost": {
                        "type": "number",
                        "format": "float",
                        "example": 15
                    },
                    "discount": {
                        "type": "number",
                        "format": "float",
                        "example": 0
                    },
                    "total": {
                        "type": "number",
                        "format": "float",
                        "example": 314.99
                    },
                    "shipping_address": {
                        "$ref": "#/components/schemas/Address"
                    },
                    "billing_address": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Address"
                            }
                        ],
                        "nullable": true,
                        "type": "object"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItem"
                        }
                    },
                    "stripe_session_id": {
                        "type": "string",
                        "example": "cs_test_a1b2c3d4",
                        "nullable": true
                    },
                    "stripe_payment_intent_id": {
                        "type": "string",
                        "example": "pi_123456789",
                        "nullable": true
                    },
                    "paid_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "cancelled_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "refunded_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "OrderItem": {
                "title": "Order Item",
                "description": "Order Item schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "order_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 10
                    },
                    "product_name": {
                        "type": "string",
                        "example": "MacBook Pro"
                    },
                    "variant_name": {
                        "type": "string",
                        "example": "Silver",
                        "nullable": true
                    },
                    "sku": {
                        "type": "string",
                        "example": "MBP-SLV",
                        "nullable": true
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 149.99
                    },
                    "total": {
                        "type": "number",
                        "format": "float",
                        "example": 299.98
                    },
                    "variant_attributes": {
                        "description": "Selected variant attributes snapshot",
                        "type": "object",
                        "example": {
                            "color": "Gold",
                            "size": "18in"
                        },
                        "nullable": true
                    },
                    "attributes": {
                        "description": "Selected variant attributes alias",
                        "type": "object",
                        "example": {
                            "color": "Gold",
                            "size": "18in"
                        },
                        "nullable": true
                    },
                    "variant": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 2
                            },
                            "name": {
                                "type": "string",
                                "example": "Gold / 18in"
                            },
                            "sku": {
                                "type": "string",
                                "example": "NK-G-18"
                            },
                            "attributes": {
                                "description": "Attributes map for variant",
                                "type": "object",
                                "example": {
                                    "color": "Gold",
                                    "size": "18in"
                                }
                            },
                            "options": {
                                "description": "Options map alias for variant",
                                "type": "object",
                                "example": {
                                    "color": "Gold",
                                    "size": "18in"
                                }
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "NotFoundResponse": {
                "description": "Not found error response",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/NotFoundResponse"
                        }
                    }
                }
            },
...
            "UnauthorizedResponse": {
                "description": "Unauthorized error response",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/UnauthorizedResponse"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}